apps/01-wikipedia-rag-langchain/pyproject.toml
[project]
name = "01-wikipedia-rag-llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface>=0.0.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.0",
]
